clean, rm gallery dep
