+++ doc/requirements-doc.txt
@@ -39,7 +39,6 @@ sphinx==4.3.2
 sphinx-click==3.0.2
 sphinx-copybutton==0.4.0
 sphinxemoji==0.2.0
-sphinx-gallery==0.10.0
 sphinx-jsonschema==1.17.2
 sphinx-panels==0.6.0
 sphinx-version-warning==1.1.2
